# Perf optimize interleaved weighted round robin scheduler