--- a/common/tasks/interleaved_weighted_round_robin.go
+++ b/common/tasks/interleaved_weighted_round_robin.go
@@ -55,6 +55,7 @@ type (
 		notifyChan   chan struct{}
 		shutdownChan chan struct{}
 
+		numInflightTask int64
 		sync.RWMutex
 		priorityToWeight     map[Priority]int
 		weightToTaskChannels map[int]*WeightedChannel
@@ -89,6 +90,7 @@ func NewInterleavedWeightedRoundRobinScheduler(
 		notifyChan:   make(chan struct{}, 1),
 		shutdownChan: make(chan struct{}),
 
+		numInflightTask:      0,
 		priorityToWeight:     option.PriorityToWeight,
 		weightToTaskChannels: make(map[int]*WeightedChannel),
 		iwrrChannels:         iwrrChannels,
@@ -134,6 +136,7 @@ func (s *InterleavedWeightedRoundRobinScheduler) Submit(
 ) {
 	channel := s.getOrCreateTaskChannel(s.priorityToWeight[task.GetPriority()])
 	channel.Chan() <- task
+	atomic.AddInt64(&s.numInflightTask, 1)
 	s.notifyDispatcher()
 }
 
@@ -143,6 +146,7 @@ func (s *InterleavedWeightedRoundRobinScheduler) TrySubmit(
 	channel := s.getOrCreateTaskChannel(s.priorityToWeight[task.GetPriority()])
 	select {
 	case channel.Chan() <- task:
+		atomic.AddInt64(&s.numInflightTask, 1)
 		s.notifyDispatcher()
 		return true
 	default:
@@ -238,18 +242,12 @@ LoopDispatch:
 			continue LoopDispatch
 		}
 	}
+	atomic.AddInt64(&s.numInflightTask, -numTasks)
 }
 
 func (s *InterleavedWeightedRoundRobinScheduler) hasRemainingTasks() bool {
-	s.RLock()
-	defer s.RUnlock()
-
-	for _, weightedChan := range s.weightToTaskChannels {
-		if weightedChan.Len() > 0 {
-			return true
-		}
-	}
-	return false
+	numTasks := atomic.LoadInt64(&s.numInflightTask)
+	return numTasks > 0
 }
 
 func (s *InterleavedWeightedRoundRobinScheduler) rescheduleTasks() {
@@ -269,6 +267,7 @@ DrainLoop:
 			}
 		}
 	}
+	atomic.AddInt64(&s.numInflightTask, -numTasks)
 }
 
 func (s *InterleavedWeightedRoundRobinScheduler) isStopped() bool {