--- a/common/tasks/interleaved_weighted_round_robin_test.go
+++ b/common/tasks/interleaved_weighted_round_robin_test.go
@@ -27,7 +27,6 @@ package tasks
 import (
 	"math/rand"
 	"sync"
-	"sync/atomic"
 	"testing"
 
 	"github.com/golang/mock/gomock"
@@ -107,7 +106,6 @@ func (s *interleavedWeightedRoundRobinSchedulerSuite) TestSubmitSchedule_Success
 	s.scheduler.Submit(mockTask)
 
 	testWaitGroup.Wait()
-	s.Equal(atomic.LoadInt64(&s.scheduler.numInflightTask), int64(0))
 }
 
 func (s *interleavedWeightedRoundRobinSchedulerSuite) TestSubmitSchedule_Fail() {
@@ -133,7 +131,6 @@ func (s *interleavedWeightedRoundRobinSchedulerSuite) TestSubmitSchedule_Fail() 
 	s.scheduler.Submit(mockTask)
 
 	testWaitGroup.Wait()
-	s.Equal(atomic.LoadInt64(&s.scheduler.numInflightTask), int64(0))
 }
 
 func (s *interleavedWeightedRoundRobinSchedulerSuite) TestChannels() {
@@ -229,5 +226,4 @@ func (s *interleavedWeightedRoundRobinSchedulerSuite) TestParallelSubmitSchedule
 	endWaitGroup.Wait()
 
 	testWaitGroup.Wait()
-	s.Equal(atomic.LoadInt64(&s.scheduler.numInflightTask), int64(0))
 }