<!-- Describe what has changed in this PR -->
**What changed?**
* Perf optimize interleaved weighted round robin scheduler
  * Bypass scheduler if there is no other task
  * Remove unnecessary shutdown check

<!-- Tell your future self why have you made these changes -->
**Why?**
```bash
before
BenchmarkInterleavedWeightedRoundRobinScheduler_Sequential-16    	 6610477	       176.0 ns/op	       8 B/op	       1 allocs/op
BenchmarkInterleavedWeightedRoundRobinScheduler_Parallel-16    	 3230086	       346.7 ns/op	       8 B/op	       1 allocs/op

after
BenchmarkInterleavedWeightedRoundRobinScheduler_Sequential-16    	26722878	        46.48 ns/op	       8 B/op	       1 allocs/op
BenchmarkInterleavedWeightedRoundRobinScheduler_Parallel-16    	 4259040	       267.4 ns/op	       8 B/op	       1 allocs/op
```

<!-- How have you verified this change? Tested locally? Added a unit test? Checked in staging env? -->
**How did you test it?**
Updated tests & benchmark

<!-- Assuming the worst case, what can be broken when deploying this change to production? -->
**Potential risks**
N/A

<!-- Is this PR a hotfix candidate or require that a notification be sent to the broader community? (Yes/No) -->
**Is hotfix candidate?**
N/A